+++ react-carousel/my-app/src/App.tsx
@@ -2,10 +2,42 @@ import { Carousel } from './Carousel ';
 import './App.css';
 import './index.css';
 
+export type Image = {
+  src: string;
+  alt: string;
+};
+
+const images: Image[] = [
+  {
+    src: '/images/fushiguro.webp',
+    alt: 'Megumi Fushiguro',
+  },
+  {
+    src: '/images/inumaki.webp',
+    alt: 'Toge Inumaki',
+  },
+  {
+    src: '/images/itadori.webp',
+    alt: 'Yuji Itadori',
+  },
+  {
+    src: '/images/kugisaki.webp',
+    alt: 'Nobara Kugisaki',
+  },
+  {
+    src: '/images/panda.webp',
+    alt: 'Panda',
+  },
+  {
+    src: '/images/zen-in.webp',
+    alt: "Maki Zen'in",
+  },
+];
+
 export default function App() {
   return (
     <>
-      <Carousel />
+      <Carousel images={images} />
     </>
   );
 }
